% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rashomonlite.R
\name{grow_from}
\alias{grow_from}
\title{Expand top models to the next dimension}
\usage{
grow_from(best_df, all_vars, k_next, m)
}
\arguments{
\item{best_df}{A data frame of best models from the previous dimension (as returned by [select_best()]).}

\item{all_vars}{Character vector of all available predictor names.}

\item{k_next}{Integer, size of the next dimension of models to grow to.}

\item{m}{Integer specifying the maximum number of new models to generate.}
}
\value{
A list of unique character vectors, where each vector represents the set
  of predictors for a candidate model of dimension `k_next`.
}
\description{
Expand top models to the next dimension
}
\author{
Prince Mensah Ansah,
        Mark Castilo Philip,
        Mohammad Al Srayheen
}
