% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rashomonlite.R
\name{select_best}
\alias{select_best}
\title{Select the top-performing models based on AIC}
\usage{
select_best(models_df, alpha = 0.5)
}
\arguments{
\item{models_df}{A data frame returned by `fit_models_dim()`, containing AIC values.}

\item{alpha}{Numeric between 0 and 1 specifying the fraction of models to retain.}
}
\value{
A subset of the input data frame containing only the top `alpha` proportion
  of models sorted by lowest AIC.
}
\description{
This function takes a data frame of models (as returned by `fit_models_dim()`)
and returns the top fraction `alpha` of models sorted by lowest AIC.
}
\author{
Prince Mensah Ansah,
        Mark Castilo Philip,
        Mohammad Al Srayheen
}
